requirements.txt
beautifulsoup4==4.9.1
Django==2.2.12
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-guardian==2.2.0
django-js-asset==1.2.2
django-mptt==0.11.0
Faker==4.1.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
